+++ src/nodejs/lambda-layers/database-util/src/query/index.js
@@ -9,7 +9,7 @@ const parameterize = (query, params) => {
   const text = query.replace(/\{\{(.*?)\}\}/g, (match, token) => {
     const value = token.split('.')
       .reduce((obj, key) => key === '$' ? obj : obj[key]===null ? null : obj[key], params);
-    values.push(pubCleanString(value));
+    values.push(value);
     return `$${values.length}`;
   });
   return { text, values };