+++ terraform/variables.tf
@@ -99,3 +99,7 @@ variable "meditor_service_username" {
 variable "meditor_service_password" {
   type = string
 }
+
+variable "environment" {
+  type = string
+}